telegram-api/requirements.txt
requests==2.25.1
python-telegram-bot==13.6
requests==2.25.1
python-telegram-bot==13.6
werkzeug==0.16.1
flask==1.1.1
gunicorn==20.1.0
flask-restplus==0.13.0
webargs==6.1.0